hw6/requirements.txt
asgiref~=3.3.1
Django~=3.1.4
djangorestframework~=3.12.2
pytz~=2019.3
sqlparse==0.4.1

celery~=5.0.5
six~=1.14.0
websocket-client~=0.57.0
pip~=20.3.3
wheel~=0.34.2
PyYAML~=5.3.1
numpy~=1.19.4
setuptools~=45.2.0
requests~=2.22.0
PyJWT~=1.7.1
cryptography~=2.8
configparser~=5.0.1
lxml~=4.6.2
keyring~=18.0.1
amqp~=5.0.2
vine~=5.0.0
colorama~=0.4.3
simplejson~=3.16.0
kombu~=5.0.2
billiard~=3.6.3.0
redis~=3.5.3
Click~=7.0
docker~=4.4.1
paramiko~=2.6.0
httplib2~=0.14.0
certifi~=2019.11.28
chardet~=3.0.4
urllib3~=1.25.8
wcwidth~=0.2.5
oauthlib~=3.1.0
blinker~=1.4
idna~=2.8
future~=0.18.2
oauth~=1.0.1
bcrypt~=3.1.7